Introduce module powerauth-fido2-model
