+++ powerauth-fido2-model/src/main/java/com/wultra/security/powerauth/fido2/model/entity/AuthenticatorParameters.java
@@ -16,8 +16,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.security.powerauth.client.model.entity.fido2;
+package com.wultra.security.powerauth.fido2.model.entity;
 
+import com.wultra.security.powerauth.fido2.model.response.AuthenticatorAttestationResponse;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
