+++ powerauth-fido2-model/src/main/java/com/wultra/security/powerauth/fido2/model/error/PowerAuthError.java
@@ -1,6 +1,6 @@
 /*
  * PowerAuth Server and related software components
- * Copyright (C) 2023 Wultra s.r.o.
+ * Copyright (C) 2018 Wultra s.r.o.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
@@ -15,22 +15,39 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+package com.wultra.security.powerauth.fido2.model.error;
 
-package com.wultra.security.powerauth.client.model.request.fido2;
+import lombok.Getter;
+import lombok.Setter;
 
-import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
+import java.io.Serial;
+import java.io.Serializable;
 
 /**
- * Request object for registration challenge.
+ * Class representing a PowerAuth error.
  *
- * @author Roman Strobl, roman.strobl@wultra.com
+ * @author Petr Dvorak, petr@wultra.com
  */
-@Data
-public class RegistrationChallengeRequest {
+@Getter
+@Setter
+public class PowerAuthError implements Serializable {
 
-    @NotBlank
-    private String userId;
-    @NotBlank
-    private String applicationId;
+    @Serial
+    private static final long serialVersionUID = 3172664961204150558L;
+
+    /**
+     * Error code.
+     */
+    private String code;
+
+    /**
+     *  Message (not localized).
+     */
+    private String message;
+
+    /**
+     * Localized message.
+     */
+    private String localizedMessage;
+
 }
