+++ powerauth-fido2-model/src/main/java/com/wultra/security/powerauth/fido2/model/request/AssertionChallengeRequest.java
@@ -16,29 +16,30 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.security.powerauth.client.model.entity.fido2;
+package com.wultra.security.powerauth.fido2.model.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
-import lombok.ToString;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
+ * Request for obtaining assertion challenge.
+ *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class AuthenticatorAssertionResponse {
+public class AssertionChallengeRequest {
 
-    @NotBlank
-    private String clientDataJSON;
-
-    @NotBlank
-    private String authenticatorData;
-
+    private String userId;
     @NotEmpty
-    @ToString.Exclude
-    private byte[] signature;
-
-    private String userHandle;
+    private List<@NotBlank String> applicationIds;
+    private String externalId;
+    @NotBlank
+    private String templateName;
+    private Map<String, String> parameters = new HashMap<>();
 
 }
