+++ powerauth-fido2-model/src/main/java/com/wultra/security/powerauth/fido2/model/request/AssertionVerificationRequest.java
@@ -16,11 +16,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.security.powerauth.client.model.request.fido2;
+package com.wultra.security.powerauth.fido2.model.request;
 
 import com.fasterxml.jackson.annotation.JsonSetter;
 import com.fasterxml.jackson.annotation.Nulls;
-import com.wultra.security.powerauth.client.model.entity.fido2.AuthenticatorAssertionResponse;
+import com.wultra.security.powerauth.fido2.model.response.AuthenticatorAssertionResponse;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
