+++ powerauth-fido2-model/src/main/java/com/wultra/security/powerauth/fido2/model/request/RegisteredAuthenticatorsRequest.java
@@ -16,37 +16,22 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.security.powerauth.client.model.entity.fido2;
+package com.wultra.security.powerauth.fido2.model.request;
 
-import com.wultra.security.powerauth.client.model.enumeration.ActivationStatus;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
 /**
- * Information about a registered authenticator.
+ * Request for obtaining list of registered authenticators for given user.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class AuthenticatorDetail {
+public class RegisteredAuthenticatorsRequest {
 
+    @NotBlank
     private String userId;
-    private String activationId;
+    @NotBlank
     private String applicationId;
-    private String activationName;
-    private String credentialId;
-    private ActivationStatus activationStatus;
-    private Map<String, Object> extras;
-    private String platform;
-    private String deviceInfo;
-    private String blockedReason;
-    private long failedAttempts;
-    private long maxFailedAttempts;
-    private List<String> applicationRoles = new ArrayList<>();
-    private List<String> activationFlags = new ArrayList<>();
-    private byte[] publicKeyBytes;
 
 }
