+++ powerauth-fido2-model/src/main/java/com/wultra/security/powerauth/fido2/model/response/AssertionVerificationResponse.java
@@ -16,35 +16,30 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.security.powerauth.client.model.response.fido2;
+package com.wultra.security.powerauth.fido2.model.response;
 
-import com.wultra.security.powerauth.client.model.enumeration.ActivationStatus;
+import com.wultra.security.powerauth.fido2.model.enumeration.ActivationStatus;
 import lombok.Data;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 /**
+ * Response for the assertion verification.
+ *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class RegistrationResponse {
+public class AssertionVerificationResponse {
 
+    private boolean assertionValid;
     private String userId;
     private String activationId;
     private String applicationId;
-    private String credentialId;
-    private String activationName;
     private ActivationStatus activationStatus;
-    private Map<String, Object> extras;
-    private String platform;
-    private String deviceInfo;
     private String blockedReason;
-    private long failedAttempts;
-    private long maxFailedAttempts;
+    private long remainingAttempts;
     private List<String> applicationRoles = new ArrayList<>();
     private List<String> activationFlags = new ArrayList<>();
-    private byte[] publicKeyBytes;
 
 }
