+++ powerauth-fido2-model/src/main/java/com/wultra/security/powerauth/fido2/model/response/AuthenticatorAssertionResponse.java
@@ -1,6 +1,6 @@
 /*
  * PowerAuth Server and related software components
- * Copyright (C) 2021 Wultra s.r.o.
+ * Copyright (C) 2023 Wultra s.r.o.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
@@ -16,28 +16,29 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.security.powerauth.client.model.response.fido2;
+package com.wultra.security.powerauth.fido2.model.response;
 
-import com.wultra.security.powerauth.client.model.entity.fido2.AllowCredentials;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
 import lombok.ToString;
 
-import java.util.List;
-
 /**
- * Response for obtaining assertion challenge.
- *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class AssertionChallengeResponse {
+public class AuthenticatorAssertionResponse {
 
-    private List<String> applicationIds;
+    @NotBlank
+    private String clientDataJSON;
+
+    @NotBlank
+    private String authenticatorData;
+
+    @NotEmpty
     @ToString.Exclude
-    private String challenge;
-    private String userId;
-    private Long failedAttempts;
-    private Long maxFailedAttempts;
-    private List<AllowCredentials> allowCredentials;
+    private byte[] signature;
+
+    private String userHandle;
 
 }
