+++ powerauth-fido2/pom.xml
@@ -69,6 +69,10 @@
             <groupId>io.getlime.security</groupId>
             <artifactId>powerauth-client-model</artifactId>
         </dependency>
+        <dependency>
+            <groupId>io.getlime.security</groupId>
+            <artifactId>powerauth-fido2-model</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>org.bouncycastle</groupId>
