+++ powerauth-rest-client-spring/src/main/java/com/wultra/security/powerauth/rest/client/PowerAuthFido2RestClient.java
@@ -24,13 +24,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.wultra.core.rest.client.base.DefaultRestClient;
 import com.wultra.core.rest.client.base.RestClient;
 import com.wultra.core.rest.client.base.RestClientException;
-import com.wultra.security.powerauth.client.PowerAuthFido2Client;
-import com.wultra.security.powerauth.client.model.entity.fido2.AuthenticatorAssertionResponse;
-import com.wultra.security.powerauth.client.model.entity.fido2.AuthenticatorParameters;
-import com.wultra.security.powerauth.client.model.error.PowerAuthClientException;
-import com.wultra.security.powerauth.client.model.error.PowerAuthError;
-import com.wultra.security.powerauth.client.model.request.fido2.*;
-import com.wultra.security.powerauth.client.model.response.fido2.*;
+import com.wultra.security.powerauth.fido2.client.PowerAuthFido2Client;
+import com.wultra.security.powerauth.fido2.model.entity.AuthenticatorParameters;
+import com.wultra.security.powerauth.fido2.model.error.PowerAuthClientException;
+import com.wultra.security.powerauth.fido2.model.error.PowerAuthError;
+import com.wultra.security.powerauth.fido2.model.request.*;
+import com.wultra.security.powerauth.fido2.model.response.*;
 import io.getlime.core.rest.model.base.request.ObjectRequest;
 import io.getlime.core.rest.model.base.response.ObjectResponse;
 import org.slf4j.Logger;
